requirements.txt
black==23.3.0
click==8.1.3
execnet==2.0.1
iniconfig==2.0.0
mock==5.0.2
mypy-extensions==1.0.0
packaging==23.1
paho-mqtt==1.6.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pytest==7.4.0
